Messing around with pluto animations

+++ Assets/Scripts/Scriptable-Objects/Pluto.cs
@@ -4,8 +4,8 @@ using UnityEngine;
 [CreateAssetMenu(menuName = "ScriptableObjects/GameObjects/Enemys/Pluto")]
 public class Pluto : BasicAttributes
 {
-    public int moveInterval;
-    public int cerberusCooldown;
+    public float moveInterval;
+    public float cerberusCooldown;
     public float cerberusRange;
     public float cerberusSpeed;
     public float statueDelay;
